Extract throttler logic into common function
